+++ docs/changelog.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-* :feature:`292` Add informatics param to LiquidHandle
+* :feature:`292` Add informatics param to p.transfer
 * :support:`291` Update copyright and authors
 * :feature:`290` Add informatics attribute to Instruction